fix: honor durable snapshot warning threshold
