+++ pkg/services/factory_runtime/internal/services/orchestration/runtime/captured_turn_terminate_integration_test.go
@@ -11,9 +11,11 @@ import (
 
 	"go.uber.org/zap"
 
+	"github.com/portpowered/infinite-you/internal/testutil/recordingfixtures"
 	platformclock "github.com/portpowered/infinite-you/pkg/platform/clock"
 	"github.com/portpowered/infinite-you/pkg/platform/logging"
 	eventswire "github.com/portpowered/infinite-you/pkg/services/events/wire"
+	interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
 	factoryruntime "github.com/portpowered/infinite-you/pkg/services/factory_runtime"
 	factorysessions "github.com/portpowered/infinite-you/pkg/services/factory_sessions"
 	factorysessionswire "github.com/portpowered/infinite-you/pkg/services/factory_sessions/wire"
@@ -23,6 +25,91 @@ import (
 	workersessionswire "github.com/portpowered/infinite-you/pkg/services/worker_sessions/wire"
 	"github.com/portpowered/infinite-you/pkg/services/workers"
 )
+
+func TestRecordedWorkerSessionObservationReprojectsWhenRestoredHistoryGrows(t *testing.T) {
+	base := time.Date(2026, 8, 8, 12, 0, 0, 0, time.UTC)
+	workID := "work-restored-growth"
+	events := recordedObservationTestEvents(t, base, workID)
+	prefix := append([]interfaces.FactoryEvent(nil), events...)
+	events = append(events, interfaces.FactoryEvent{
+		Context: interfaces.FactoryEventContext{
+			Tick: 6, Sequence: 1, EventTime: base.Add(6 * time.Second),
+			WorkIDs: stringSliceForRecordedTest([]string{workID}),
+		},
+		Type: interfaces.FactoryEventTypeWorkStateChange,
+	})
+	restored := interfaces.FactoryWorldState{
+		WorkItemsByID: map[string]work.FactoryWorkItem{workID: {ID: workID}},
+		CompletedDispatches: []interfaces.FactoryWorldDispatchCompletion{
+			{
+				DispatchID:  "dispatch-early",
+				StartedAt:   base.Add(time.Second),
+				CompletedAt: base.Add(3 * time.Second),
+				WorkItemIDs: []string{workID},
+				Result:      interfaces.WorkstationResult{Outcome: string(workers.OutcomeAccepted)},
+			},
+			{
+				DispatchID:  "dispatch-late",
+				StartedAt:   base.Add(5 * time.Second),
+				CompletedAt: base.Add(7 * time.Second),
+				WorkItemIDs: []string{workID},
+				Result:      interfaces.WorkstationResult{Outcome: string(workers.OutcomeAccepted)},
+			},
+		},
+	}
+
+	t.Run("projects appended runtime event", func(t *testing.T) {
+		projectorCalls := 0
+		service := newRecordedWorkerSessionObservationWithRestoredState(
+			nil,
+			&recordingfixtures.ScriptedRuntimeLedger{Events: events},
+			func(_ []interfaces.FactoryEvent, _ int) (interfaces.FactoryWorldState, error) {
+				projectorCalls++
+				return restored, nil
+			},
+			platformclock.Real{},
+			nil,
+			nil,
+			"",
+			nil,
+			&restored,
+			prefix,
+		)
+
+		result, err := service.ListObservations(context.Background(), workersessions.ListObservationsRequest{WorkID: workID})
+		if err != nil {
+			t.Fatalf("ListObservations() error = %v", err)
+		}
+		if projectorCalls != 1 {
+			t.Fatalf("full world projection calls = %d, want 1 after appended runtime event", projectorCalls)
+		}
+		if len(result.Observations) != 2 || result.Observations[0].State != workersessions.StateCompleted || result.Observations[1].State != workersessions.StateCompleted {
+			t.Fatalf("reprojected observations = %#v, want two projected completed attempts", result.Observations)
+		}
+	})
+
+	t.Run("propagates projection failure", func(t *testing.T) {
+		service := newRecordedWorkerSessionObservationWithRestoredState(
+			nil,
+			&recordingfixtures.ScriptedRuntimeLedger{Events: events},
+			func(_ []interfaces.FactoryEvent, _ int) (interfaces.FactoryWorldState, error) {
+				return interfaces.FactoryWorldState{}, errors.New("projection failed")
+			},
+			platformclock.Real{},
+			nil,
+			nil,
+			"",
+			nil,
+			&restored,
+			prefix,
+		)
+
+		_, err := service.ListObservations(context.Background(), workersessions.ListObservationsRequest{WorkID: workID})
+		if !errors.Is(err, workersessions.ErrObservationProjectionUnavailable) {
+			t.Fatalf("ListObservations() error = %v, want projection unavailable", err)
+		}
+	})
+}
 
 // TestTerminateFactorySession_FansOutCapturedChildrenBeforeTargetCleanup
 // exercises the committed Factory Sessions close boundary against the real
